refactor(*)!: move start functionality to wadm lib
