+++ src/main.rs
@@ -1,48 +1,13 @@
 use core::net::SocketAddr;
 
 use std::path::PathBuf;
-use std::sync::Arc;
-use std::time::Duration;
 
 use anyhow::Context as _;
-use async_nats::jetstream::{stream::Stream, Context};
-use bytes::Bytes;
 use clap::Parser;
-use hyper_util::rt::{TokioExecutor, TokioIo};
-use nats::StreamPersistence;
-use tokio::net::TcpListener;
-use tokio::spawn;
-use tokio::sync::Semaphore;
-use tracing::{debug, error};
+use wadm::{config::WadmConfig, start_wadm, StreamPersistence};
 use wadm_types::api::DEFAULT_WADM_TOPIC_PREFIX;
 
-use wadm::{
-    consumers::{
-        manager::{ConsumerManager, WorkerCreator},
-        *,
-    },
-    nats_utils::LatticeIdParser,
-    scaler::manager::{ScalerManager, WADM_NOTIFY_PREFIX},
-    server::{ManifestNotifier, Server},
-    storage::{nats_kv::NatsKvStore, reaper::Reaper},
-    workers::{CommandPublisher, CommandWorker, EventWorker, StatusPublisher},
-    DEFAULT_COMMANDS_TOPIC, DEFAULT_EVENTS_TOPIC, DEFAULT_MULTITENANT_EVENTS_TOPIC,
-    DEFAULT_STATUS_TOPIC, DEFAULT_WADM_EVENTS_TOPIC, DEFAULT_WADM_EVENT_CONSUMER_TOPIC,
-};
-
-mod connections;
 mod logging;
-mod nats;
-mod observer;
-
-use connections::ControlClientConstructor;
-
-const WADM_EVENT_STREAM_NAME: &str = "wadm_events";
-const WADM_EVENT_CONSUMER_STREAM_NAME: &str = "wadm_event_consumer";
-const COMMAND_STREAM_NAME: &str = "wadm_commands";
-const STATUS_STREAM_NAME: &str = "wadm_status";
-const NOTIFY_STREAM_NAME: &str = "wadm_notify";
-const WASMBUS_EVENT_STREAM_NAME: &str = "wasmbus_events";
 
 #[derive(Parser, Debug)]
 #[command(name = clap::crate_name!(), version = clap::crate_version!(), about = "wasmCloud Application Deployment Manager", long_about = None)]
@@ -251,6 +216,37 @@ struct Args {
     http_admin: Option<SocketAddr>,
 }
 
+impl From<Args> for WadmConfig {
+    fn from(config: Args) -> WadmConfig {
+        WadmConfig {
+            host_id: config.host_id,
+            domain: config.domain,
+            max_jobs: config.max_jobs,
+            nats_server: config.nats_server,
+            nats_seed: config.nats_seed,
+            nats_jwt: config.nats_jwt,
+            nats_creds: config.nats_creds,
+            nats_tls_ca_file: config.nats_tls_ca_file,
+            state_bucket: config.state_bucket,
+            cleanup_interval: config.cleanup_interval,
+            api_prefix: config.api_prefix,
+            stream_prefix: config.stream_prefix,
+            manifest_bucket: config.manifest_bucket,
+            multitenant: config.multitenant,
+            max_state_bucket_bytes: config.max_state_bucket_bytes,
+            max_manifest_bucket_bytes: config.max_manifest_bucket_bytes,
+            stream_persistence: config.stream_persistence,
+            max_command_stream_bytes: config.max_command_stream_bytes,
+            max_event_stream_bytes: config.max_event_stream_bytes,
+            max_event_consumer_stream_bytes: config.max_event_consumer_stream_bytes,
+            max_status_stream_bytes: config.max_status_stream_bytes,
+            max_notify_stream_bytes: config.max_notify_stream_bytes,
+            max_wasmbus_event_stream_bytes: config.max_wasmbus_event_stream_bytes,
+            http_admin: config.http_admin,
+        }
+    }
+}
+
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     let args = Args::parse();
@@ -258,352 +254,32 @@ async fn main() -> anyhow::Result<()> {
     logging::configure_tracing(
         args.structured_logging,
         args.tracing_enabled,
-        args.tracing_endpoint,
+        args.tracing_endpoint.clone(),
     );
 
-    // Build storage adapter for lattice state (on by default)
-    let (client, context) = nats::get_client_and_context(
-        args.nats_server.clone(),
-        args.domain.clone(),
-        args.nats_seed.clone(),
-        args.nats_jwt.clone(),
-        args.nats_creds.clone(),
-        args.nats_tls_ca_file.clone(),
-    )
-    .await?;
-
-    // TODO: We will probably need to set up all the flags (like lattice prefix and topic prefix) down the line
-    let connection_pool = ControlClientConstructor::new(client.clone(), None);
-
-    let trimmer: &[_] = &['.', '>', '*'];
-
-    let store =
-        nats::ensure_kv_bucket(&context, args.state_bucket, 1, args.max_state_bucket_bytes).await?;
-
-    let state_storage = NatsKvStore::new(store);
-
-    let manifest_storage = nats::ensure_kv_bucket(
-        &context,
-        args.manifest_bucket,
-        1,
-        args.max_manifest_bucket_bytes,
-    )
-    .await?;
-
-    let internal_stream_name = |stream_name: &str| -> String {
-        match args.stream_prefix.clone() {
-            Some(stream_prefix) => {
-                format!(
-                    "{}.{}",
-                    stream_prefix.trim_end_matches(trimmer),
-                    stream_name
-                )
-            }
-            None => stream_name.to_string(),
-        }
-    };
-
-    debug!("Ensuring wadm event stream");
-
-    let event_stream = nats::ensure_limits_stream(
-        &context,
-        internal_stream_name(WADM_EVENT_STREAM_NAME),
-        vec![DEFAULT_WADM_EVENTS_TOPIC.to_owned()],
-        Some(
-            "A stream that stores all events coming in on the wadm.evt subject in a cluster"
-                .to_string(),
-        ),
-        args.max_event_stream_bytes,
-        args.stream_persistence.into(),
-    )
-    .await?;
-
-    debug!("Ensuring command stream");
-
-    let command_stream = nats::ensure_stream(
-        &context,
-        internal_stream_name(COMMAND_STREAM_NAME),
-        vec![DEFAULT_COMMANDS_TOPIC.to_owned()],
-        Some("A stream that stores all commands for wadm".to_string()),
-        args.max_command_stream_bytes,
-        args.stream_persistence.into(),
-    )
-    .await?;
-
-    let status_stream = nats::ensure_status_stream(
-        &context,
-        internal_stream_name(STATUS_STREAM_NAME),
-        vec![DEFAULT_STATUS_TOPIC.to_owned()],
-        args.max_status_stream_bytes,
-        args.stream_persistence.into(),
-    )
-    .await?;
-
-    debug!("Ensuring wasmbus event stream");
-
-    // Remove the previous wadm_(multitenant)_mirror streams so that they don't
-    // prevent us from creating the new wasmbus_(multitenant)_events stream
-    // TODO(joonas): Remove this some time in the future once we're confident
-    // enough that there are no more wadm_(multitenant)_mirror streams around.
-    for mirror_stream_name in &["wadm_mirror", "wadm_multitenant_mirror"] {
-        if (context.get_stream(mirror_stream_name).await).is_ok() {
-            context.delete_stream(mirror_stream_name).await?;
-        }
-    }
-
-    let wasmbus_event_subjects = match args.multitenant {
-        true => vec![DEFAULT_MULTITENANT_EVENTS_TOPIC.to_owned()],
-        false => vec![DEFAULT_EVENTS_TOPIC.to_owned()],
-    };
-
-    let wasmbus_event_stream = nats::ensure_limits_stream(
-        &context,
-        WASMBUS_EVENT_STREAM_NAME.to_string(),
-        wasmbus_event_subjects.clone(),
-        Some(
-            "A stream that stores all events coming in on the wasmbus.evt subject in a cluster"
-                .to_string(),
-        ),
-        args.max_wasmbus_event_stream_bytes,
-        args.stream_persistence.into(),
-    )
-    .await?;
-
-    debug!("Ensuring notify stream");
-
-    let notify_stream = nats::ensure_notify_stream(
-        &context,
-        NOTIFY_STREAM_NAME.to_owned(),
-        vec![format!("{WADM_NOTIFY_PREFIX}.*")],
-        args.max_notify_stream_bytes,
-        args.stream_persistence.into(),
-    )
-    .await?;
-
-    debug!("Ensuring event consumer stream");
-
-    let event_consumer_stream = nats::ensure_event_consumer_stream(
-        &context,
-        WADM_EVENT_CONSUMER_STREAM_NAME.to_owned(),
-        DEFAULT_WADM_EVENT_CONSUMER_TOPIC.to_owned(),
-        vec![&wasmbus_event_stream, &event_stream],
-        Some(
-            "A stream that sources from wadm_events and wasmbus_events for wadm event consumer's use"
-                .to_string(),
-        ),
-        args.max_event_consumer_stream_bytes,
-        args.stream_persistence.into(),
-    )
-    .await?;
-
-    debug!("Creating event consumer manager");
-
-    let permit_pool = Arc::new(Semaphore::new(
-        args.max_jobs.unwrap_or(Semaphore::MAX_PERMITS),
-    ));
-    let event_worker_creator = EventWorkerCreator {
-        state_store: state_storage.clone(),
-        manifest_store: manifest_storage.clone(),
-        pool: connection_pool.clone(),
-        command_topic_prefix: DEFAULT_COMMANDS_TOPIC.trim_matches(trimmer).to_owned(),
-        publisher: context.clone(),
-        notify_stream,
-        status_stream: status_stream.clone(),
-    };
-    let events_manager: ConsumerManager<EventConsumer> = ConsumerManager::new(
-        permit_pool.clone(),
-        event_consumer_stream,
-        event_worker_creator.clone(),
-        args.multitenant,
-    )
-    .await;
-
-    debug!("Creating command consumer manager");
-
-    let command_worker_creator = CommandWorkerCreator {
-        pool: connection_pool,
-    };
-    let commands_manager: ConsumerManager<CommandConsumer> = ConsumerManager::new(
-        permit_pool.clone(),
-        command_stream,
-        command_worker_creator.clone(),
-        args.multitenant,
-    )
-    .await;
-
-    // TODO(thomastaylor312): We might want to figure out how not to run this globally. Doing a
-    // synthetic event sent to the stream could be nice, but all the wadm processes would still fire
-    // off that tick, resulting in multiple people handling. We could maybe get it to work with the
-    // right duplicate window, but we have no idea when each process could fire a tick. Worst case
-    // scenario right now is that multiple fire simultaneously and a few of them just delete nothing
-    let reaper = Reaper::new(
-        state_storage.clone(),
-        Duration::from_secs(args.cleanup_interval / 2),
-        [],
-    );
-
-    let wadm_event_prefix = DEFAULT_WADM_EVENTS_TOPIC.trim_matches(trimmer);
-
-    debug!("Creating lattice observer");
-
-    let observer = observer::Observer {
-        parser: LatticeIdParser::new("wasmbus", args.multitenant),
-        command_manager: commands_manager,
-        event_manager: events_manager,
-        reaper,
-        client: client.clone(),
-        command_worker_creator,
-        event_worker_creator,
-    };
-
-    debug!("Subscribing to API topic");
-
-    let server = Server::new(
-        manifest_storage,
-        client,
-        Some(&args.api_prefix),
-        args.multitenant,
-        status_stream,
-        ManifestNotifier::new(wadm_event_prefix, context),
-    )
-    .await?;
-
-    if let Some(addr) = args.http_admin {
-        let socket = TcpListener::bind(addr)
-            .await
-            .context("failed to bind on HTTP administation endpoint")?;
-        let svc = hyper::service::service_fn(move |req| {
-            const OK: &str = r#"{"status":"ok"}"#;
-            async move {
-                let (http::request::Parts { method, uri, .. }, _) = req.into_parts();
-                match (method.as_str(), uri.path()) {
-                    ("HEAD", "/livez") => Ok(http::Response::default()),
-                    ("GET", "/livez") => Ok(http::Response::new(http_body_util::Full::new(
-                        Bytes::from(OK),
-                    ))),
-                    (method, "/livez") => http::Response::builder()
-                        .status(http::StatusCode::METHOD_NOT_ALLOWED)
-                        .body(http_body_util::Full::new(Bytes::from(format!(
-                            "method `{method}` not supported for path `/livez`"
-                        )))),
-                    ("HEAD", "/readyz") => Ok(http::Response::default()),
-                    ("GET", "/readyz") => Ok(http::Response::new(http_body_util::Full::new(
-                        Bytes::from(OK),
-                    ))),
-                    (method, "/readyz") => http::Response::builder()
-                        .status(http::StatusCode::METHOD_NOT_ALLOWED)
-                        .body(http_body_util::Full::new(Bytes::from(format!(
-                            "method `{method}` not supported for path `/readyz`"
-                        )))),
-                    (.., path) => http::Response::builder()
-                        .status(http::StatusCode::NOT_FOUND)
-                        .body(http_body_util::Full::new(Bytes::from(format!(
-                            "unknown endpoint `{path}`"
-                        )))),
-                }
-            }
-        });
-        let srv = hyper_util::server::conn::auto::Builder::new(TokioExecutor::new());
-        spawn(async move {
-            loop {
-                let stream = match socket.accept().await {
-                    Ok((stream, _)) => stream,
-                    Err(err) => {
-                        error!(?err, "failed to accept HTTP administration connection");
-                        continue;
-                    }
-                };
-                let svc = svc.clone();
-                if let Err(err) = srv.serve_connection(TokioIo::new(stream), svc).await {
-                    error!(?err, "failed to serve HTTP administration connection");
-                }
-            }
-        });
-    }
-
+    let mut wadm = start_wadm(args.into())
+        .await
+        .context("failed to run wadm")?;
     tokio::select! {
-        res = server.serve() => {
-            res?
+        res = wadm.join_next() => {
+            match res {
+                Some(Ok(_)) => {
+                    tracing::info!("WADM has exited successfully");
+                    std::process::exit(0);
+                }
+                Some(Err(e)) => {
+                    tracing::error!("WADM has exited with an error: {:?}", e);
+                    std::process::exit(1);
+                }
+                None => {
+                    tracing::info!("WADM server did not start");
+                    std::process::exit(0);
+                }
+            }
         }
-        res = observer.observe(wasmbus_event_subjects) => {
-            res?
+        _ = tokio::signal::ctrl_c() => {
+            tracing::info!("Received Ctrl+C, shutting down");
+            std::process::exit(0);
         }
-        _ = tokio::signal::ctrl_c() => {}
-    }
-    Ok(())
-}
-
-#[derive(Clone)]
-struct CommandWorkerCreator {
-    pool: ControlClientConstructor,
-}
-
-#[async_trait::async_trait]
-impl WorkerCreator for CommandWorkerCreator {
-    type Output = CommandWorker;
-
-    async fn create(
-        &self,
-        lattice_id: &str,
-        multitenant_prefix: Option<&str>,
-    ) -> anyhow::Result<Self::Output> {
-        let client = self.pool.get_connection(lattice_id, multitenant_prefix);
-
-        Ok(CommandWorker::new(client))
-    }
-}
-
-#[derive(Clone)]
-struct EventWorkerCreator<StateStore> {
-    state_store: StateStore,
-    manifest_store: async_nats::jetstream::kv::Store,
-    pool: ControlClientConstructor,
-    command_topic_prefix: String,
-    publisher: Context,
-    notify_stream: Stream,
-    status_stream: Stream,
-}
-
-#[async_trait::async_trait]
-impl<StateStore> WorkerCreator for EventWorkerCreator<StateStore>
-where
-    StateStore: wadm::storage::Store + Send + Sync + Clone + 'static,
-{
-    type Output = EventWorker<StateStore, wasmcloud_control_interface::Client, Context>;
-
-    async fn create(
-        &self,
-        lattice_id: &str,
-        multitenant_prefix: Option<&str>,
-    ) -> anyhow::Result<Self::Output> {
-        let client = self.pool.get_connection(lattice_id, multitenant_prefix);
-        let command_publisher = CommandPublisher::new(
-            self.publisher.clone(),
-            &format!("{}.{lattice_id}", self.command_topic_prefix),
-        );
-        let status_publisher = StatusPublisher::new(
-            self.publisher.clone(),
-            Some(self.status_stream.clone()),
-            &format!("wadm.status.{lattice_id}"),
-        );
-        let manager = ScalerManager::new(
-            self.publisher.clone(),
-            self.notify_stream.clone(),
-            lattice_id,
-            multitenant_prefix,
-            self.state_store.clone(),
-            self.manifest_store.clone(),
-            command_publisher.clone(),
-            status_publisher.clone(),
-            client.clone(),
-        )
-        .await?;
-        Ok(EventWorker::new(
-            self.state_store.clone(),
-            client,
-            command_publisher,
-            status_publisher,
-            manager,
-        ))
     }
 }
